+++ CHANGELOG.md
@@ -28,7 +28,7 @@ Fixed
   * AZ-263: Update `ip_restriction` block with new attributes
 
 Added
-  * AZ-374: Add option to create or not the application insight
+  * AZ-374: Add option to create or not the application insights
   
 # v3.1.0 - 2020-07-30