+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-# unreleased
+# v4.2.0 - 2021-03-09
 
 Added
   * AZ-424: Add Service Tags (AzureRM v2.42+)
-  * AZ-374: Allow disabling Application Insights deployment.
+  * AZ-374: Allow disabling Application Insights deployment
 
 # v4.1.0 - 2021-01-15